services/assistant/requirements-dev.txt
-r requirements.txt
alembic==1.13.3
ruff==0.7.0
pytest-asyncio==0.24.0
grpcio-tools==1.67.0
